requirements.txt
Flask==3.0.2
Flask-WTF==1.2.1
gunicorn==20.0.4
python-dotenv==1.0.1